feat: enable ppr
