+++ package.json
@@ -18,7 +18,7 @@
     "kysely": "^0.27.1",
     "kysely-planetscale": "^1.4.0",
     "medium-zoom": "^1.1.0",
-    "next": "14.0.4",
+    "next": "^14.0.5-canary.49",
     "next-themes": "0.2.1",
     "react": "^18.2.0",
     "react-compare-image": "^3.4.0",
