+++ src/app/blog/[slug]/page.tsx
@@ -1,13 +1,10 @@
 import type { Metadata } from 'next';
 import { notFound, redirect } from 'next/navigation';
-import { Suspense } from 'react';
 
 import { Mdx } from '@/components/views/blog/mdx';
 import { createMetadata } from '@/utils/metadata';
 import { buildOgImageUrl } from '@/utils/og';
 import { allBlogs } from 'contentlayer/generated';
-
-import Loading from '../../loading';
 
 import type { BlogPostPageContext } from './types';
 
@@ -17,12 +14,7 @@ export default function BlogPostPage(context: BlogPostPageContext) {
 
   if (!slug || !post) return notFound();
   if (post.link) return redirect(post.link);
-
-  return (
-    <Suspense fallback={<Loading />}>
-      <Mdx code={post.body.code} />
-    </Suspense>
-  );
+  return <Mdx code={post.body.code} />;
 }
 
 export const generateStaticParams = () =>
