+++ src/components/views/blog/reactions/use-reactions.tsx
@@ -3,7 +3,8 @@
 import confetti from 'canvas-confetti';
 import { useCallback, useEffect, useState, type MouseEvent } from 'react';
 
-import { getReactions, incrementReaction } from '@/actions/reactions';
+import { incrementReaction } from '@/db/actions';
+import { getReactions } from '@/db/queries';
 import { useHasMounted } from '@/hooks/use-has-mounted';
 import { useWindowDimensions } from '@/hooks/use-window-dimensions';
 import type { ReactionName, ReactionsCounters } from '@/lib/planetscale';
