+++ src/db/actions.ts
@@ -1,13 +1,7 @@
-'use server';
-
 import { unstable_noStore as noStore } from 'next/cache';
 import { cache } from 'react';
 
-import {
-  db,
-  type ReactionName,
-  type ReactionsCounters,
-} from '@/lib/planetscale';
+import { db, type ReactionName } from '@/lib/planetscale';
 
 import { canRunAction } from './utils';
 
@@ -34,24 +28,20 @@ export const incrementReaction = cache(
   },
 );
 
-export const getReactions = cache(
-  async (slug: string): Promise<ReactionsCounters> => {
-    try {
-      const data = await db
-        .selectFrom('counters')
-        .where('slug', '=', slug)
-        .select(['likes', 'loves', 'awards', 'bookmarks'])
-        .execute();
-      const [counters] = data;
-      if (!counters) return {};
-      return {
-        likes: counters.likes || 0,
-        loves: counters.loves || 0,
-        awards: counters.awards || 0,
-        bookmarks: counters.bookmarks || 0,
-      };
-    } catch (e) {
-      return {};
-    }
-  },
-);
+export const recordView = cache(async (slug: string) => {
+  if (!canRunAction) return;
+  noStore();
+  try {
+    const data = await db
+      .selectFrom('counters')
+      .where('slug', '=', slug)
+      .select(['views'])
+      .execute();
+
+    const views = !data.length ? 0 : Number(data[0].views);
+    db.insertInto('counters')
+      .values({ slug, views: 1 })
+      .onDuplicateKeyUpdate({ views: views + 1 })
+      .execute();
+  } catch (e) {}
+});
